pyproject.toml
[tool.poetry]
name = "quotes"
version = "0.1.0"
description = ""
authors = ["Revuka Oleksandr <a.revuka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
psycopg2-binary = "^2.9.7"
mongoengine = "^0.27.0"
pymongo = "^4.5.0"
sqlalchemy = "^2.0.20"
pillow = "^10.0.0"
python-dotenv = "^1.0.1"
django-environ = "^0.11.2"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 130
